third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-transform-list-single-values-expected.txt

This is a testharness.js-based test.
[FAIL] Animating a custom property of type <transform-list> containing a single value with iterationComposite
  assert_equals: expected "translateX(250px)" but got "translate(250px)"
Harness: the test ran to completion.
